requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.6
Django==4.0.4
django-braces==1.15.0
django-crispy-forms==1.14.0
django-ipware==4.0.2
django-model-utils==4.2.0
django-star-ratings==0.9.2
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
sorl-thumbnail==12.8.0
sqlparse==0.4.2
swapper==1.3.0
tzdata==2022.1
